Add files via upload
In this program, the user is asked to enter 20 integers. If the array is not sorted, it will be sorted using the Quick Sort algorithm. Then, the user is prompted to enter a number to search. The program will then perform both binary search and linear search on the sorted array (if necessary) and display the results along with the time taken for each search.

Note that Quick Sort has an average-case time complexity of O(n log n), which is better than O(n^2) algorithms like Bubble Sort or Selection Sort. This makes Quick Sort efficient for sorting large arrays. Additionally, Binary Search has a time complexity of O(log n) compared to Linear Search's O(n) time complexity, which makes it much faster for large sorted arrays.